refactor facet list to use publishValue
